fix(Transaction): Fixed one wrong assert and added resetting sessionID when no calls.
Fixed one assert which still had the old logic of vector in its check. Also
added new function for dialog removal which checks if this was the last dialog,
then resets the sessionID counter to 1. This is a quick fix for asserts that
activate with larger sessionID:s after sessions have been terminated. This
does not address the issue and the issue should be addressed by replacing
all vectors with maps for sessionID.