refactor(Initiation): Renamed SDP functions and separated SDP from Transactions.
The SIP Manager is now responsible for all the interactions between SIP components.
The SIP manager now gets SDP directly from GlobalSDPState instead of trough
transactions. GlobalSDPState is now the definitive holder of SDP messages. This
was needed so that conference call SDP can be easily compiled in GlobalSDPState.

Broken: Calling not working.