refactor(Initiation): Redistribute some SIPDialog responsibilities elsewhere.
May also fix some issues related to connection type. The generation of branch
parameter is now the responsibility of client. The setting of correct address
and port for via and contact fields is not the responsibility of SIPTransport.
Created a constant variable for setting the connection type to reduce the number
of errors created by unset connection type and for easier future change.