refactor(Negotiation): Moved STUN server request to NetworkCandidates.
The Stun class had two separate responsibilities (binding and server request)
which had very little to do with each other and if you used them at the same
time it would break. The previous implementation did not use them at the same
time, but now the code is clearer and easier to modify.