fix(Initiation): Remove session from negotiation if other person ends the call.
There was a clear situation where the other person was not removed from negotiation
if the other person happened to close the call. This problem manifested itself when
someone tried to call a second time and they were assigned the same sessionID.
The results was an invalid SDP in INVITE Ok response.