fix(Transaction): Fixed not ending call. BYE timeout now ends call.
The bug was following: After receiving a BYE we ended the call and sent OK
response. The problem was that the ending of the call would last so long, that
BYE sender transaction timed out resulting resulting in sender not recognizing the
response when it arrived. THis reverses the order so ok is sent before call des-
truction and BYE timeout also ends the call.