refactor(Transport): Moved message sanity checks to their own module
Parsing was getting too big and it will probably include more parsing stuff
in the future so a division was necessary. Having the message checks in their
own module makes sense. The interface only includes two functions, but the
implementation is over 500 lines. SIPTransport is also getting big and this
helps a little bit on that as well.