feature(Transport): Accept-field composition and improved content composition
Had to modify the composing and parsing pipelines to handle empty fields as
accept-field can have empty.

At the same time made some improvements to content conversions in siptransport
and conversions. The flow of content conversions is currently not suitable for
non text content, which should be fixed at some point.