feature(Transport): Parsing of Authentication fields
Paring of Authorization, Proxy-Authenticate, Proxy-Authorization and
WWW-Authenticate fields. Implemented parsing using map where all the Digest
values are stored. Modified Authentication-Info to use that same method.

Getting authentication to work also required small change to fieldparsing and
the URI parsing for not including <> as part of the URI but as part of
name-addr.