refactor(Transaction): Separate SIPSingleCall from SIPClient
This gets rid of SIP Clients dependencies on sessionID and transaction user
which make no sense since registration is also using SIP Client.

This commit also removes the separate BYE timeout signal which now means that
if bye timeouts the call will not end. The current implementation of BYE
timeout would have caused a HEAP corruption which are difficult to debug so
a better solution should be devised later.