refactor(Transport): Replace transportID with remote address
The transportID was a needles layer when we could just as well always use the
remote address to find the correct connection. All the messages and dialogs
have the information what is the local address the dialog/message belongs to.

One architectural problem did result from removal of transport id since we
don't known the remote address with new received connections and have to wait
needlessly for the connection to be initialized. This will hopefully be removed
at some point in these architecture changes.