refactor(Initiation): Dialog architecture upgrade
This commit makes the user dialog creation and SIP Transaction message
processing use the new architecture of message flow. The commit also includes
small fixes.

What works:
- Starting the call

What does not:
- Ending the call

The problem with ending the call is caused when calling ourselves the BYEs are
sent at the same time and both are discarded because of invalid CSeq.

List of updates:
- Message flow class
- Constructing the dialog using flow class
- Storing flow parts in new message flow class
- Setting local address for negotiation during startup. This address has to be
there, but has no purpose when ICE is in use.
- Modify the flow so that registration/sipsinglecall also use processor class.
Registration does not yet have fully use the new architecture.
- Modify ICE logic in controller so that ICE success delivers the SDP:s so
there is no need to query them afterwards. This logic also facilitates the
call success and ice success to arrive at different times.