refactor(Transport): Implement the new architecture for SIP Transport
Really like the new architecture. Everything seems to be working fine with
light testing. Now also the outgoing transport flow is using the processor and
flow classes. This enables easy addition of new processor modules.

In transport one has to take care that the flow goes in other direction
compared to registration and dialog. Since the dialog and transport are now
completely decoupled, the message should find its way where it is going
whatever the address is.

There might be some room to improve the determination of correct from-address
for incoming dialogs. The current code is somewhat complicated and relies on
data from the caller.