feature(Transaction): Localhost calls even if we use location service
This commit enables calling localhost calls even if we are registered to
server. The previous logic sent everything to our proxy, but there are
situations when the proxy cannot reach the said destination.

The logic could be improved to handle also LAN addresses, but that does not
seem important enough at the moment. If we are not registered, Kvazzup
functions normally as peer-to-peer application and requires firewall to be open
for incoming TCP connection on port 5060.