feature(Negotiation): Run-time STUN server adjustment
Went a bit overboard with this one. User can now switch the stun server or
media ports while the call is running! Tested it and it seems to work now.
This won't of course affect the current call, but future calls will (or not)
use the STUN address provided by the user. Haven't actually test