fix(Processing): Fix hangs and crashes with pulseaudio on Linux
The QAudioOutput is now stopped at destructor so that it does not crash
occasionally.

For some reason, the QAudioInput suspend hangs sometimes. A working fix
in this commit was to switch suspends to stops on Linux. This probably
takes more time than suspend, but it no longer seems to hang.