feature(Processing): Reduce audio buffers to reduce latency
On windows, the mixing buffer doesn't have to be so large. On both linux and
windows, it is not nice if the output buffer starts filling up so it is also
reduced. There is an issue on linux where the outputfilter is not able to take
data fast enough, causing the buffer to fill.