feature(Processing): Also do AGC at output and mix no longer clips
Made the DSP filter configurable and this enabled easy implementation of
an additional AGC before audio output. The mixing is now a sum that is divided
by the number of inputs. This would have previously resulted in quiet audio if
only one is speaking.

This theory has not been thoroughly tested. More audio testing is needed to
verify that this all makes sense.