feature(Processing): Read everything available from mic
Before we read the periodsize as many times as possible, but that didn't work
very well on Linux. This commit makes it so we read everything available at
once.

This also increased the size of audio samples for linux since they seemed to
overflow from the buffers with smaller samples due to large amount of samples
available at the input at once.

There are still some missing samples at output on Linux. I'm suspecting this
is caused by how the samples are copied at the output.