feature(Processing): Buffer AEC playback frames
The previously included audioframebuffer works wonderfully as an playback frame
buffer. Now the echo input sounds correct and we can easily buffer it a little
bit so it is inputted at the right time to AEC.

For the first time, AEC seems to actually do something. The optimal echo delay
length is tricky to estimate though.