refactor(UI): Divide CallWindow into UIManager and CallWindow
This division adds necessary division for different ways of user interacting
with Kvazzup. Most prominently the addition of tray many which is completely
separate from the main menu becomes much easier when the call window is not
responsible for everything. The new UI manager can always decide which way
the user interface reacts to users actions and whether it involves the main
window or not.