refactor(Initiation): Distribute SIPRegistration tasks elsewhere
The SIP Registration didn't have clear enough responsibilities to
justify its existence. Its responsibilities are distributed to SIPClient
and SIPManager and the class is removed in this update.

One small change in this commit to functionality is that the periodic
re-REGISTER is done in always, not just after successful registrations.