fix(Processing): Make it clear what is the current orientation of video
There were weird orientation bugs hanging around because one for the
conversions flips the video which counteracted the flipping of the
camera on windows. This has been a cause for confusion for a long time
in Kvazzup. This does not fix the conversions, but at least it flips it
to correct orientation so no other filters have to suffer.