refactor(GUI): Remove viewID/videoID from VideoviewFactory
The original idea was that there could be more than one destination
for video from peer, but this does not seem likely at this point.
Even if both screen sharing and the video would be combined, they
would still likely be combined into one view inside the VideoWidget.

Removed the possibility of multiple views for one peer to remove
unnecessary complexity from code.