feature(Processing): Measure decoding latency more accurately and reduce it
There is no longer any frames put needlessly into a buffer. The slice
implementation does not work, nor did it previously, but there were
structures for it for no reason.

The frame info (called Data) is now put into a buffer that is used
to measure the latency more accurately.