refactor(Global): Move ICE from initiation to media
After getting more familiar with the subject, I've come to the conclusion
that ICE should only be performed after the call has been negotiated on the ISP
side. This also makes more sense from the architecture point of view.

The ICE will now be performed as a part of delivery, right before creating
media. There is less need for complicated delivery of SDP messages for ICE
in initiation.