fix(GUI): Fix crash with fullscreen when call ends
The fix type was confusing in that the software was exiting when the call
ended, which might indicate some other problems. In any case this commit
fixes this particular crash (which is the only one I'm aware of as of this
moment) regarding the fullscreen being detached when the call ends.

The solution here is to simply reattach it before removing the participant.