feature(Media): Update project to Qt6
Updated Kvazzup to Qt version 6 because the CMake build system is better
supported in this version. The update was pretty straight forward. The only
problem I have found is related to Motion Jpeg camera format, which is actually
supported fully in the new version, but this causes the problem in Kvazzup,
since Kvazzup relies on the invalid conversion to RGB32 which happened in
background of previous version of Qt.

I haven't tested all performance characteristics of this new version.

I haven't tested Linux version, support for that comes later.