build(CMake): Build uvgRTP, openHEVC on Linux, Crypto++ and opus dependencies
These projects either use CMake or have a CMake available making them the
easiest to build from CMake. I did encounter linking error for m.lib when the
opus shared lib was enabled, but currently this error does not show up for
some reason.

Currently, OpenHEVC does not build with MSVC and was left out. I haven't
tested Linux in a while, but that is on my todo list.