feature(Processing): Greatly increase the number of supported camera formats
This integrates the LibYUV based conversion filter to Kvazzup, increasing the
number of supported camera input formats from 4 to 13.

Some of the previous conversion filters are now unused and will probably be
removed soon. Libyuv mostly provides conversions to yuv420p, but does not
provide conversion to RGB formats used by Qt. This means that the existing
YUV to RGB conversion filter is still required.