build(Project): Update project file to latest configuration
I was already going to abandon the old qmake file, but I couldn't get the
deployment working with CMake so I tried to deploy with QMake. Unfortunately
there is some sort of incompatibility with Qt6 that I don't have time or
interest to solve. I made this commit to update the project file in case I
want to try this again.