fix(Settings): Remove linux special case for camera formats
In Qt5, the camera on Linux did not seem to work unless given a specific
format. The camera format seems to work somewhat well in Qt6 and therefore the
special case is no longer needed. There are some minor differences between
operating systems though that cause some confusion for me.