fix(Processing): Fix MJPEG format in Qt before 6.5
Qt 6.5 adopts FFmpeg as the default backend for camera input. This produces
actual MJPEG input, but Qt versions before 6.5 are still producing RGB32.
This commit gets Kvazzup to support both scenarios.

Qt version 6.5 and up would probably need some sort of condition to see if
lib jpeg has been included so the conversion for MJPEG actually works, but this
is outside the scope of this commit.