fix(Settings): Fix UI values of bit rates when less than 3 digits
The bit rate setting values slightly above full mbits, were shown incorrectly
because they weren't padded with zeros as they are supposed to be. This commit
fixes this issues.

I also set the rounding for bit rates to 10 kbit for video because adjusting
individual kbits seemed to muddy the UI more than it is worth.