fix(Settings): Avoid selecting low/high fps values ever
The debug mode defaults to low complexity for ease of use, but this lead
to it selecting very weird camera options such as 2 fps of over FullHD
which did not even work correctly. This commit sets the points values for
low and high fps values to 0 so they (almost) never get selected.