fix(Negotiation): Dont assume SSRC has been generated when there is only one
This commit fixes calling ourselves by not assuming our other SSRC as being
generated. When SSRC is generated for us by host, there must be two SSRCs in
the message, one for the peer, and one for us. When calling ourselves there is
one from ourselves and no other SSRCs.