refactor(Global): Use SSRCs instead of MediaID to track streams
Previously we could only track media using one of the SSRCs, but with SFU,
there needs to be a way to track remote and local media separately dues to
the possibility of one side having different count of ssrcs in their media
which represent media streams. SSRCs alone are the most logical choice.

This does meant that from now on SSRCs will be mandatory in each SDP message,
but this should not pose an issue as they would be recommended in any case.