feature(Delivery): Clients send start stop commands inside RTCP APP
Whenever we decide that we should switch to P2P link, we need to synchronize
the stopping with the SFU so that other participants down links are not
overloaded. This commit delays the switching to P2P link and vice versa so
that the SFU can make the switch at the same time.