fix(Processing): Mark updating hybrid indexes as unimplemented and stop using
The index and rtp pointer that the mediamanager provides is only for the newest
one and this is not correct when we want to update the indexes of existing p2p
links.

This should fix case where uses are only added, but updating existing indexes
is needed for removal to work.